Refactor project structure
Refactor project structure into engine library and ECSTestGame executable.
The library contains all reusable components whereas the executable
constains only application-specific code. Directory structure is modified
to accomodate the library-executable split.
