+++ include/engine/CMakeLists.txt
@@ -1,0 +1,7 @@
+# Add headers
+file(GLOB SUB_HEADERS "*.hpp")
+
+set(ENGINE_HEADERS
+    ${ENGINE_HEADERS}
+    ${SUB_HEADERS}
+    PARENT_SCOPE)
