+++ include/game/CollisionEvent.hpp
