+++ include/game/EventHandlers.hpp
@@ -7,8 +7,8 @@
 
 
 #include <vector>
+#include <engine/EventHandler.hpp>
 
-#include "EventHandler.hpp"
 #include "CollisionEvent.hpp"
 #include "Logics.hpp"
