+++ include/game/SpriteComponent.hpp
