--- src/EventTypeId.cpp
@@ -1,8 +1,0 @@
-//
-// Created by Lehdari on 8.1.2019.
-//
-
-#include "EventTypeId.hpp"
-
-
-int64_t EventTypeId::__typeIdCounter = 0;
