+++ src/engine/EventComponent.cpp
@@ -2,7 +2,7 @@
 // Created by Lehdari on 11.4.2019.
 //
 
-#include "EventComponent.hpp"
+#include <engine/EventComponent.hpp>
 
 
 DEFINE_COMPONENT_TEMPLATES(EventComponent);
