+++ src/engine/EventTypeId.cpp
@@ -1,0 +1,8 @@
+//
+// Created by Lehdari on 8.1.2019.
+//
+
+#include <engine/EventTypeId.hpp>
+
+
+int64_t EventTypeId::__typeIdCounter = 0;
