+++ src/game/PhysicsComponent.cpp
@@ -2,7 +2,7 @@
 // Created by Lehdari on 11.4.2019.
 //
 
-#include "PhysicsComponent.hpp"
+#include <game/PhysicsComponent.hpp>
 
 
 DEFINE_COMPONENT_TEMPLATES(PhysicsComponent);
