+++ src/game/PhysicsSystem.cpp
@@ -2,7 +2,7 @@
 // Created by lehdari on 3.11.2018.
 //
 
-#include "PhysicsSystem.hpp"
+#include <game/PhysicsSystem.hpp>
 
 
 void PhysicsSystem::operator()(const EntityId& eId, PhysicsComponent& c)
