+++ src/game/SpriteComponent.cpp
@@ -2,7 +2,7 @@
 // Created by Lehdari on 11.4.2019.
 //
 
-#include "SpriteComponent.hpp"
+#include <game/SpriteComponent.hpp>
 
 
 DEFINE_COMPONENT_TEMPLATES(SpriteComponent);
